+++ setup.cfg
@@ -34,9 +34,8 @@ include_package_data = True
 setup_requires =
   setuptools_scm
 install_requires =
-  pfsspy>=1.1.2
   pyvista[all]>= 0.38.4
-  sunpy[map]>=5.0.0,<5.1.0
+  sunpy[map]>=5.0.0
 
 [options.extras_require]
 tests =
@@ -44,6 +43,7 @@ tests =
   pytest-cov
   pytest-doctestplus
   scikit-image
+  sunkit-magex
 docs =
   ipython
   sphinx
