+++ sunkit_pyvista/tests/test_plotting.py
@@ -4,17 +4,11 @@ This file contains figure comparison tests.
 
 import astropy.constants as const
 import astropy.units as u
-import matplotlib.pyplot as plt
-import numpy as np
-import pfsspy
 import pytest
 import pyvista
 import sunpy.data.test as test
 import sunpy.map as smap
 from astropy.coordinates import SkyCoord
-from matplotlib import colors
-from pfsspy import tracing
-from pfsspy.sample_data import get_gong_map
 from sunpy.coordinates import frames
 
 from sunkit_pyvista import SunpyPlotter
@@ -77,30 +71,4 @@ def test_plot_map_with_functionality(
     plotter = SunpyPlotter()
     plotter.load(filepath)
 
-    plotter.show(cpos=(0, 1, 0), before_close_callback=verify_cache_image)
-
-
-def test_field_lines_figure(aia171_test_map, plotter, verify_cache_image):
-    gong_fname = get_gong_map()
-    gong_map = smap.Map(gong_fname)
-    nrho = 35
-    rss = 2.5
-    lat = np.linspace(-np.pi / 2, np.pi / 2, 8, endpoint=False)
-    lon = np.linspace(0, 2 * np.pi, 8, endpoint=False)
-    lat, lon = np.meshgrid(lat, lon, indexing="ij")
-    lat, lon = lat.ravel() * u.rad, lon.ravel() * u.rad
-    radius = 1.2
-    tracer = tracing.PythonTracer()
-    input_ = pfsspy.Input(gong_map, nrho, rss)
-    output_ = pfsspy.pfss(input_)
-    seeds = SkyCoord(lon, lat, radius * const.R_sun, frame=gong_map.coordinate_frame)
-    field_lines = tracer.trace(seeds, output_)
-
-    def color_function(field_line):
-        norm = colors.LogNorm(vmin=1, vmax=1000)
-        cmap = plt.get_cmap("magma")
-        return cmap(norm(np.abs(field_line.expansion_factor)))
-
-    plotter.plot_map(aia171_test_map)
-    plotter.plot_field_lines(field_lines, color_func=color_function)
     plotter.show(cpos=(0, 1, 0), before_close_callback=verify_cache_image)
